update cleanup and refactoring
