+++ src/main/java/org/ietf/jose/jwk/key/EllipticCurveJwk.java
@@ -15,10 +15,8 @@
  */
 package org.ietf.jose.jwk.key;
 
+import lombok.Data;
 import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
 import org.ietf.jose.adapter.XmlAdapterBigIntegerBase64Url;
 import org.ietf.jose.jwk.JWK;
 
@@ -52,9 +50,7 @@ import java.math.BigInteger;
  * @author Key Bridge
  */
 @EqualsAndHashCode(callSuper = true)
-@ToString
-@Getter
-@Setter
+@Data
 @XmlAccessorType(XmlAccessType.FIELD)
 public class EllipticCurveJwk extends JWK {
