+++ src/main/java/org/ietf/jose/jws/JwsHeader.java
@@ -15,10 +15,8 @@
  */
 package org.ietf.jose.jws;
 
+import lombok.Data;
 import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.Setter;
-import lombok.ToString;
 import org.ietf.jose.jwk.JWK;
 
 import java.net.URI;
@@ -46,9 +44,7 @@ import java.util.List;
  * @since 0.0.1 created 02/01/2018
  */
 @EqualsAndHashCode(callSuper = true)
-@ToString
-@Getter
-@Setter
+@Data
 public class JwsHeader extends AbstractHeader {
 
   /**
